+++ lib/cake_flutter.dart
@@ -4,7 +4,7 @@ import 'dart:async';
 
 import 'package:cake/cake.dart';
 import 'package:flutter/gestures.dart';
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 export 'package:cake/cake.dart';
 
@@ -14,6 +14,7 @@ part 'flutter_test_options.dart';
 part 'flutter_test_runner.dart';
 part 'index_options.dart';
 part 'search.dart';
+part 'test_element_actions.dart';
 part 'test_element_wrapper_collection.dart';
 part 'test_element_wrapper.dart';
 part 'widget_tree.dart';