+++ lib/index_options.dart
@@ -9,10 +9,21 @@ class IndexOptions {
   /// Must have debug tree flag on to work
   final List<WidgetType> debugTypeFilter;
 
-  /// Only index these types and their children. Recommended to help with performance
+  /// Only index these types and their children.
+  ///
+  /// Note [indexKeys] ignores types that do not match.
+  /// Recommended to help with performance since searching will be faster.
+  /// However double check that this list is correct as it will cause tests to
+  /// fail if trying to search for a non-indexed type.
+  final List<WidgetType> indexTypes;
+
+  /// Only index these widgets and their children. Ignores [indexTypes] if
+  /// the type does not match.
+  ///
+  /// Recommended to help with performance
   /// since searching will be faster. However double check that this list is correct
   /// as it will cause tests to fail if trying to search for a non-indexed type.
-  final List<WidgetType> indexFilter;
+  final List<Key> indexKeys;
 
   /// Helper for common search types for Text and Icons
   static const List<WidgetType> defaultSearchTypes = [
@@ -20,18 +31,48 @@ class IndexOptions {
     WidgetType<Icon>(),
   ];
 
+  /// By default, Error Widgets will throw when found. This can be disabled
+  /// to allow the search to continue.
+  final bool warnOnErrorWidgets;
+
   IndexOptions({
     this.debugTree = false,
     this.debugTypeFilter = const [],
-    this.indexFilter = const [],
+    this.indexTypes = const [],
+    this.indexKeys = const [],
+    this.warnOnErrorWidgets = true,
   });
 
-  bool matchesIndexFilter(Element element) {
-    if (indexFilter.isEmpty) {
-      return true;
-    } else {
-      return indexFilter.any((filterElement) => filterElement.match(element));
+  void checkForErrorWidgets(Element element) {
+    if (element.widget is ErrorWidget && warnOnErrorWidgets) {
+      final String widgetMessage = (element.widget as ErrorWidget).message;
+      throw 'Found an ErrorWidget with the following message. To ignore this and continue, set IndexOptions.warnOnErrorWidgets to false:\n$widgetMessage';
     }
+  }
+
+  bool matchesIndexFilter(Element element) {
+    if (indexTypes.isEmpty && indexKeys.isEmpty) {
+      return true;
+    }
+
+    // Make sure to record any ErrorWidgets
+    if (warnOnErrorWidgets && element.widget is ErrorWidget) {
+      return true;
+    }
+
+    if (indexKeys.isNotEmpty) {
+      final bool keyMatch =
+          indexKeys.any((filterKey) => filterKey == element.widget.key);
+      if (keyMatch) {
+        return true;
+      }
+    }
+
+    if (indexTypes.isNotEmpty) {
+      return indexTypes.any((filterElement) => filterElement.match(element));
+    }
+
+    return false;
   }
 
   bool matchesDebugFilter(Element element) {
@@ -48,10 +89,10 @@ class IndexOptions {
   }
 
   bool hasIndexedType<W2 extends Widget>() {
-    if (indexFilter.isEmpty) {
+    if (indexTypes.isEmpty) {
       return true;
     } else {
-      return indexFilter.any((filterElement) => filterElement.matchType<W2>());
+      return indexTypes.any((filterElement) => filterElement.matchType<W2>());
     }
   }
 }